# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/v1/tags": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取多个文章标签",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{\"lists\":[{\"id\":1,\"created_on\":0,\"modified_on\":1573561866,\"name\":\"1113111113\",\"created_by\":\"\",\"modified_by\":\"guowei\",\"state\":1}], \"total\":1}, \"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "新增文章标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "State",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "CreateBy",
                        "name": "create_by",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "10001": {
                        "description": "{\"code\":10001, \"data\":{}, \"msg\":\"已存在该标签名称\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "200": {
                        "description": "{\"code\":200, \"data\":{}, \"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/tags/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "summary": "修改文章标签",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ModifiedBy",
                        "name": "modified_by",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "10002": {
                        "description": "{\"code\":10002, \"data\":{}, \"msg\":\"该标签不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "200": {
                        "description": "{\"code\": 200, \"data\":{}, \"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "summary": "删除文章标签",
                "responses": {
                    "10002": {
                        "description": "{\"code\":10002,\"data\":{},\"msg\":\"该标签不存在\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "200": {
                        "description": "{\"code\":200, \"data\":{}, \"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}